Adding more detail to the programming model readme and exercises

+++ ManagedMemory/CPU_Code/cpu_code.c
@@ -9,16 +9,16 @@ void cpu_func(double *in_h, double *out_h, int M){
 
 int main(int argc, char *argv[]) {
    int M=100000;
-   int Msize=M*sizeof(double);
+   size_t Msize=M*sizeof(double);
    double sum=0.0;
 
-   double* in_h = (double*)malloc(Msize);
-   double* out_h = (double*)malloc(Msize);
+   double *in_h = (double*)malloc(Msize);
+   double *out_h = (double*)malloc(Msize);
 
    for (int i=0; i<M; i++) // initialize
       in_h[i] = 1.0;
 
-      cpu_func(in_h, out_h, M);
+   cpu_func(in_h, out_h, M);
 
    for (int i=0; i<M; i++) // CPU-process
      sum += out_h[i];
